requirements.txt
streamlit=1.31.1
python-dotenv=1.0.1
langchain=0.1.9